Finish porting the remaining buildscripts. Declare versions and other variables where needed to the root project etc.
Moved the plugin dependencies into the buildSrc buildscript classpath so
that we can access their classes within the buildSrc. I had not previously
known that one could not access plugins defined by the parent project
script without the plugins being exposed to the buildSrc classpath
explicitly. My guess is that the plugins are provided as dependencies for
the project's buildscript because they are already applied to the
buildscript as a "project", so access to the plugins themselves through
that classpath is required.

Merged all the dependency objects into a single file, SpongeAPI.kt to
allow the changes required to add dependencies to remain small (three
lines in a single file is easier than one line in three files).

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>